Extract config struct to config package.
Export ReadConfig
Export Config, Config.IsSink, Config.IsSource
Export NameMatcher, NameMatcher.MatchMethodName.
Extract Config.sendsToIOWriter -> levee.sendsToIOWriter as a special case.
